add aconection with api of weather

+++ src/TemplatePrism/Models/Condition.cs
@@ -12,6 +12,6 @@ namespace TemplatePrism.Models
         public string Icon { get; set; }
 
         [JsonProperty("code")]
-        public int Code { get; set; }
+        public long Code { get; set; }
     }
 }
